backend/api/views/saved_tests.py
"""
Saved Tests Views
Presentation/API Layer - Endpoints for managing reusable test definitions
"""

from rest_framework import status
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import IsAuthenticated
from rest_framework.response import Response


@api_view(["GET", "POST"])
@permission_classes([IsAuthenticated])
def saved_tests_list_create(request):
    """
    GET /saved-tests - Get all saved tests for the current user
    POST /saved-tests - Save a new reusable test definition

    GET: Returns list of all Test records owned by the authenticated user.
    POST: Creates a Test record that can be reused for multiple executions.
          Stores test steps, URL, and expected behavior.

    POST Request Body:
    - test_name (required): Name of the test
    - description (optional): Test description
    - url (required): Target URL
    - steps (required): JSON array of test steps
    - expected_behavior (optional): Expected behavior description

    Returns:
    GET:
    - 200: Array of saved test objects
    - 404: User not found

    POST:
    - 201: Full saved test object
    - 400: Missing required field
    - 404: User not found
    """
    if request.method == "GET":
        # TODO: Implement saved tests list retrieval
        # 1. Query Test filtered by user
        # 2. Order by created_at descending
        # 3. Return serialized list
        return Response([])
    else:  # POST
        # TODO: Implement saved test creation
        # 1. Validate request data
        # 2. Create Test record with user association
        # 3. Return serialized Test object
        return Response(
            {
                "id": "placeholder-id",
                "test_name": "Placeholder Test",
                "url": "",
                "steps": [],
            },
            status=status.HTTP_201_CREATED,
        )


@api_view(["GET", "PUT", "DELETE"])
@permission_classes([IsAuthenticated])
def saved_test_detail(request, test_id):
    """
    GET /saved-tests/<test_id> - Retrieve one saved test definition
    PUT /saved-tests/<test_id> - Update a saved test definition
    DELETE /saved-tests/<test_id> - Delete a saved test definition

    GET: Returns complete Test record including all steps and configuration.
    PUT: Updates any provided fields of a Test record. Only updates fields that are provided.
    DELETE: Removes the Test record from the database. Does not delete associated TestExecution or TestResult records.

    Path Params:
    - test_id: UUID of the saved test

    PUT Request Body (all optional):
    - test_name: Updated test name
    - description: Updated description
    - url: Updated target URL
    - steps: Updated JSON array of steps
    - expected_behavior: Updated expected behavior

    Returns:
    GET:
    - 200: Full saved test object
    - 404: Test or user not found

    PUT:
    - 200: Updated saved test object
    - 404: Test or user not found

    DELETE:
    - 200: Success message
    - 404: Test or user not found
    """
    if request.method == "GET":
        # TODO: Implement saved test retrieval
        # 1. Get Test by ID
        # 2. Verify user ownership
        # 3. Return serialized Test object
        return Response(
            {"id": test_id, "test_name": "Placeholder Test", "url": "", "steps": []}
        )
    elif request.method == "PUT":
        # TODO: Implement saved test update
        # 1. Get Test by ID
        # 2. Verify user ownership
        # 3. Update provided fields
        # 4. Save and return updated Test object
        return Response(
            {"id": test_id, "test_name": "Updated Test", "url": "", "steps": []}
        )
    else:  # DELETE
        # TODO: Implement saved test deletion
        # 1. Get Test by ID
        # 2. Verify user ownership
        # 3. Delete Test record
        # 4. Return success message
        return Response({"message": "Test deleted successfully"})
