backend/requirements.txt
# Django and core dependencies
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9

# Celery for background tasks and scheduling
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# Authentication
djangorestframework-simplejwt>=5.3.0

# Environment variables
python-dotenv>=1.0.0

# HTTP client for external APIs
requests>=2.31.0

# Browserbase (browser sessions for recorder; official SDK)
browserbase>=1.4.0

# Playwright (recorder connects to Browserbase via CDP)
playwright>=1.40.0

# Date/time utilities
python-dateutil>=2.8.2

# UUID utilities (built-in but ensuring compatibility)
uuid>=1.30

# Linting and formatting
black>=24.0.0
flake8>=7.0.0
isort>=5.12.0

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
